feat: migrate vespa service models to consolidated API structure
- Move vespa_types.py and document_chunking.py to services/api/v1/vespa/
- Update all imports in vespa_loader service to use new API location
- Update all test imports to use new API location
- Fix relative imports in API module to use absolute imports
- All tests passing and no type errors introduced
