+++ services/vespa_loader/pubsub_consumer.py
@@ -13,12 +13,12 @@ from abc import ABC, abstractmethod
 from datetime import datetime, timezone
 from typing import Any, Callable, Dict, List, Optional, Union
 
+from services.api.v1.vespa.vespa_types import VespaDocumentType
 from services.common.config.subscription_config import SubscriptionConfig
 from services.common.logging_config import get_logger
 from services.vespa_loader.document_factory import process_message
 from services.vespa_loader.ingest_service import ingest_document_service
 from services.vespa_loader.settings import Settings
-from services.vespa_loader.vespa_types import VespaDocumentType
 
 # PubSub types
 try:
