+++ services/vespa_loader/tests/test_ingest_service.py
@@ -9,12 +9,12 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
+from services.api.v1.vespa.vespa_types import VespaDocumentType
 from services.common.http_errors import ValidationError
 from services.vespa_loader.content_normalizer import ContentNormalizer
 from services.vespa_loader.embeddings import EmbeddingGenerator
 from services.vespa_loader.ingest_service import ingest_document_service
 from services.vespa_loader.vespa_client import VespaClient
-from services.vespa_loader.vespa_types import VespaDocumentType
 
 
 class TestIngestService:
